requirements.txt
# Core dependencies
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0

# LLM Providers (install only what you need)
anthropic==0.39.0
google-generativeai==0.8.0
openai==1.54.0

# Agent framework
langgraph==0.2.45
langchain==0.3.7
langchain-anthropic==0.2.4

# Document processing
pypdf2==3.0.1
python-docx==1.1.2
reportlab==4.2.5

# FHIR
fhir.resources==7.1.0

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
jinja2==3.1.4
python-multipart==0.0.12

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0